[feat] Restore Qwen3Attention class as a subclass of QKNormRoPEAttention
- Enhanced QKNormRoPEAttention to accept positional embedding parameters.
- Inherit QKNormRoPEAttention in Qwen3Attention, specialize it for Qwen3 model.
