+++ tensorrt_llm/_torch/models/modeling_qwen3.py
@@ -1,11 +1,13 @@
-import math
 from typing import Optional, Tuple
 
 import torch
 from torch import nn
-from transformers import PretrainedConfig, Qwen3Config
+from transformers import Qwen3Config
+
+from tensorrt_llm.functional import PositionEmbeddingType
 
 from ..attention_backend import AttentionMetadata
+from ..attention_backend.interface import PositionalEmbeddingParams, RopeParams
 from ..model_config import ModelConfig
 from ..modules.decoder_layer import DecoderLayer
 from ..modules.embedding import Embedding
@@ -18,109 +20,47 @@ from .modeling_speculative import SpecDecOneEngineForCausalLM
 from .modeling_utils import DecoderModel, register_auto_model
 
 
-# Move out from this class
-def compute_yarn_parameters(
-    config: PretrainedConfig, ) -> tuple[float, float, float, float]:
-    """
-    Refer to https://github.com/huggingface/transformers/blob/main/src/transformers/modeling_rope_utils.py#L197C1-L288C1
-    Computes the inverse frequencies with NTK scaling. Please refer to the
-    [original paper](https://huggingface.co/papers/2309.00071)
-    Args:
-        config ([`~transformers.PretrainedConfig`]):
-            The model configuration.
-    Returns:
-        factor: float, the scaling factor for the RoPE embeddings
-        low: float, the lower bound of the dimension range
-        high: float, the upper bound of the dimension range
-        attention_factor: float, the post-processing scaling factor applied to the computed cos/sin
-    """
+class Qwen3Attention(QKNormRoPEAttention):
 
-    # The config does not contain rope_scaling, which means the model is not using yarn
-    rope_scaling = getattr(config, "rope_scaling", None)
-    if rope_scaling is None:
-        return 1.0, 0, 0, 1.0
+    def __init__(
+        self,
+        model_config: ModelConfig[Qwen3Config],
+        layer_idx: Optional[int] = None,
+        fuse_qk_norm_rope: bool = True,
+    ):
+        config = model_config.pretrained_config
 
-    base = config.rope_theta
-    partial_rotary_factor = config.partial_rotary_factor if hasattr(
-        config, "partial_rotary_factor") else 1.0
-    head_dim = getattr(config, "head_dim",
-                       config.hidden_size // config.num_attention_heads)
-    dim = int(head_dim * partial_rotary_factor)
-    factor = getattr(rope_scaling, "factor", 1.0)
-    attention_factor = rope_scaling.get("attention_factor")
-    mscale = rope_scaling.get("mscale")
-    mscale_all_dim = rope_scaling.get("mscale_all_dim")
-
-    if "original_max_position_embeddings" in rope_scaling:
-        original_max_position_embeddings = rope_scaling[
-            "original_max_position_embeddings"]
-        factor = config.max_position_embeddings / original_max_position_embeddings
-    else:
-        original_max_position_embeddings = config.max_position_embeddings
-
-    def get_mscale(scale, mscale=1):
-        if scale <= 1:
-            return 1.0
-        return 0.1 * mscale * math.log(scale) + 1.0
-
-    # Sets the attention factor as suggested in the paper
-    if attention_factor is None:
-        if mscale and mscale_all_dim:
-            attention_factor = float(
-                get_mscale(factor, mscale) / get_mscale(factor, mscale_all_dim))
+        if getattr(config, "rope_scaling", None) is not None:
+            if "type" in config.rope_scaling:
+                pos_type = config.rope_scaling["type"]
+            elif "rope_type" in config.rope_scaling:
+                pos_type = config.rope_scaling["rope_type"]
+            else:
+                raise ValueError(
+                    "rope_scaling must have type or rope_type field")
+            pos_embd_params = PositionalEmbeddingParams(
+                type=PositionEmbeddingType.from_string(pos_type),
+                rope=RopeParams.from_config(config),
+            )
         else:
-            attention_factor = get_mscale(factor)
+            pos_embd_params = PositionalEmbeddingParams(
+                type=PositionEmbeddingType.rope_gpt_neox,
+                rope=RopeParams.from_config(config),
+            )
 
-    # Optional config options
-    # beta_fast/beta_slow: as suggested in the paper, default to 32/1 (correspondingly)
-    beta_fast = rope_scaling.get("beta_fast") or 32
-    beta_slow = rope_scaling.get("beta_slow") or 1
-
-    # Compute the inverse frequencies
-    def find_correction_dim(num_rotations, dim, base, max_position_embeddings):
-        """Inverse dimension formula to find the dimension based on the number of rotations"""
-        return (dim *
-                math.log(max_position_embeddings /
-                         (num_rotations * 2 * math.pi))) / (2 * math.log(base))
-
-    def find_correction_range(low_rot, high_rot, dim, base,
-                              max_position_embeddings, truncate):
-        """Find dimension range bounds based on rotations"""
-        low = find_correction_dim(low_rot, dim, base, max_position_embeddings)
-        high = find_correction_dim(high_rot, dim, base, max_position_embeddings)
-        if truncate:
-            low = math.floor(low)
-            high = math.ceil(high)
-        return max(low, 0), min(high, dim - 1)
-
-    truncate = rope_scaling.get("truncate", True)
-    low, high = find_correction_range(beta_fast, beta_slow, dim, base,
-                                      original_max_position_embeddings,
-                                      truncate)
-
-    # These parts are implemented in the fusedQKNormRopeKernel.cu
-    # # def linear_ramp_factor(min, max, dim):
-    # #     if min == max:
-    # #         max += 0.001  # Prevent singularity
-
-    # #     linear_func = (torch.arange(dim, dtype=torch.float32) - min) / (max - min)
-    # #     ramp_func = torch.clamp(linear_func, 0, 1)
-    # #     return ramp_func
-
-    # # Note on variable naming: "interpolation" comes from the original technique, where we interpolate the position IDs
-    # # to expand the possible context length. In other words, interpolation = apply scaling factor.
-    # # pos_freqs = base ** (torch.arange(0, dim, 2).to(device=device, dtype=torch.float) / dim)
-    # # inv_freq_extrapolation = 1.0 / pos_freqs
-    # # inv_freq_interpolation = 1.0 / (factor * pos_freqs)
-
-    # # # Get n-dimensional rotational scaling corrected for extrapolation
-    # # inv_freq_extrapolation_factor = 1 - linear_ramp_factor(low, high, dim // 2).to(device=device, dtype=torch.float)
-    # # inv_freq = (
-    # #     inv_freq_interpolation * (1 - inv_freq_extrapolation_factor)
-    # #     + inv_freq_extrapolation * inv_freq_extrapolation_factor
-    # # )
-    # # return inv_freq, attention_factor
-    return factor, low, high, attention_factor
+        super().__init__(
+            hidden_size=config.hidden_size,
+            num_attention_heads=config.num_attention_heads,
+            num_key_value_heads=config.num_key_value_heads,
+            max_position_embeddings=config.max_position_embeddings,
+            bias=config.attention_bias,
+            pos_embd_params=pos_embd_params,
+            fuse_qk_norm_rope=fuse_qk_norm_rope,
+            layer_idx=layer_idx,
+            dtype=config.torch_dtype,
+            dense_bias=config.attention_bias,
+            config=model_config,
+        )
 
 
 class Qwen3DecoderLayer(DecoderLayer):
@@ -133,16 +73,9 @@ class Qwen3DecoderLayer(DecoderLayer):
         super().__init__()
         self.layer_idx = layer_idx
         config = model_config.pretrained_config
-        self.self_attn = QKNormRoPEAttention(
-            hidden_size=config.hidden_size,
-            num_attention_heads=config.num_attention_heads,
-            num_key_value_heads=config.num_key_value_heads,
-            max_position_embeddings=config.max_position_embeddings,
-            bias=config.attention_bias,
+        self.self_attn = Qwen3Attention(
+            model_config,
             layer_idx=layer_idx,
-            dtype=config.torch_dtype,
-            dense_bias=config.attention_bias,
-            config=model_config,
         )
 
         self.mlp = GatedMLP(
